backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pdfplumber==0.11.7
python-docx==1.2.0
rapidfuzz==3.13.0
scikit-learn==1.7.1
numpy==2.3.2
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.10.1
email-validator==2.3.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pymongo==4.14.1
requests==2.32.5
